+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Note velocity gets rounded to the nearest integer, to prevent excessive aftertouch messages from note velocity tweens.
 
+- Long arpeggio notes don't get sustained between arpeggiator loops
+
 
 
 ## 2.3.0 (2023-06-02)